jwt 테스트

+++ src/main/java/com/wgbt/wgbtapi/service/jwt/JwtService.java
@@ -7,6 +7,6 @@ public interface JwtService {
     Map<String, Object> get(String key);
     boolean isUsable(String jwt);
 
-    Long getUserNo();
+    Integer getUserNo();
 
 }
